chore(gpui): keep grid.rs under the 800-line limit
Move clipboard_rows into editing.rs so rustfmt does not push grid.rs
over the file-size cap. Drop the unused ScrollDelta import.
